+++ src/assets/SettingsAsset.php
@@ -105,7 +105,7 @@ class SettingsAsset extends AssetBundle
     {
         $blockTypes = $field->getBlockTypes();
         $blockTypeGroups = $field->getGroups();
-        $fieldLayoutHtml = self::_renderFieldLayoutHtml();
+        $fieldLayoutHtml = Neo::$plugin->blockTypes->renderFieldLayoutDesigner(new FieldLayout(['type' => Block::class]));
 
         $jsSettings = [
             'namespace' => Craft::$app->getView()->getNamespace(),
@@ -132,13 +132,7 @@ class SettingsAsset extends AssetBundle
         $jsBlockTypes = [];
 
         foreach ($blockTypes as $blockType) {
-            $fieldLayout = $blockType->getFieldLayout();
-            $oldNamespace = $view->getNamespace();
-            $view->setNamespace('neoBlockType' . $blockType['id']);
-            $fieldLayoutHtml = self::_renderFieldLayoutHtml($fieldLayout);
-            $view->setNamespace($oldNamespace);
-
-            $jsBlockType = [
+            $jsBlockTypes[] = [
                 'id' => $blockType->id,
                 'sortOrder' => $blockType->sortOrder,
                 'name' => Craft::t('site', $blockType->name),
@@ -150,13 +144,9 @@ class SettingsAsset extends AssetBundle
                 'childBlocks' => is_string($blockType->childBlocks) ? Json::decodeIfJson($blockType->childBlocks) : $blockType->childBlocks,
                 'topLevel' => (bool)$blockType->topLevel,
                 'errors' => $blockType->getErrors(),
-                'fieldLayout' => $fieldLayout->getConfig(),
-                'fieldLayoutHtml' => $fieldLayoutHtml,
-                'fieldLayoutId' => $fieldLayout->id,
+                'fieldLayoutId' => $blockType->fieldLayoutId,
                 'groupId' => $blockType->groupId,
             ];
-
-            $jsBlockTypes[] = $jsBlockType;
         }
 
         return $jsBlockTypes;
@@ -182,24 +172,5 @@ class SettingsAsset extends AssetBundle
         }
 
         return $jsBlockTypeGroups;
-    }
-
-    /**
-     * @param FieldLayout|null $fieldLayout
-     * @return string
-     */
-    private static function _renderFieldLayoutHtml(?FieldLayout $fieldLayout = null): string
-    {
-        $view = Craft::$app->getView();
-
-        // Render the field layout designer HTML, but disregard any JavaScript it outputs, as that'll be handled by Neo
-        $view->startJsBuffer();
-        $html = $view->renderTemplate('_includes/fieldlayoutdesigner', [
-            'fieldLayout' => $fieldLayout ?? new FieldLayout(['type' => Block::class]),
-            'customizableUi' => true,
-        ]);
-        $view->clearJsBuffer();
-
-        return $html;
     }
 }